🐛 Bug Fixes & Patches
Fix potential crash
Handle more errors
Some mods page switching will crash the page
Shaderpack files updates are not identified

🏗️ Refactors
Make instance install status detect more stable and fix some glitch
Aggresively improve the download perf
Unified instance resource installation
Adjust the shader/resourcepack file event handling
Add ping detection for iceservers
Allow peer user to see the ping of the negotiate server

+++ xmcl-runtime/mod/InstanceModsService.ts
@@ -1,16 +1,16 @@
-import { CurseforgeV1Client } from '@xmcl/curseforge'
+import { CurseforgeApiError, CurseforgeV1Client } from '@xmcl/curseforge'
 import { ModrinthV2Client } from '@xmcl/modrinth'
-import { InstanceModsService as IInstanceModsService, InstallMarketOptionWithInstance, InstallModsOptions, InstanceModsServiceKey, ResourceState, LockKey, MutableState, Resource, ResourceDomain, getInstanceModStateKey } from '@xmcl/runtime-api'
+import { InstanceModsService as IInstanceModsService, InstallMarketOptionWithInstance, InstallModsOptions, InstanceModsServiceKey, LockKey, Resource, ResourceDomain, ResourceState, SharedState, getInstanceModStateKey } from '@xmcl/runtime-api'
 import { emptyDir, ensureDir, rename, stat, unlink } from 'fs-extra'
 import { basename, dirname, join } from 'path'
 import { Inject, LauncherAppKey } from '~/app'
+import { InstanceService } from '~/instance'
 import { kMarketProvider } from '~/market'
 import { ResourceManager, kResourceWorker } from '~/resource'
 import { AbstractService, ExposeServiceKey, ServiceStateManager } from '~/service'
-import { AnyError, isSystemError } from '~/util/error'
+import { AnyError } from '~/util/error'
 import { LauncherApp } from '../app/LauncherApp'
-import { linkWithTimeoutOrCopy, readdirIfPresent } from '../util/fs'
-import { InstanceService } from '~/instance'
+import { linkDirectory, linkWithTimeoutOrCopy, readdirIfPresent } from '../util/fs'
 
 /**
  * Provide the abilities to import mods and resource packs files to instance
@@ -25,7 +25,7 @@ export class InstanceModsService extends AbstractService implements IInstanceMod
 
   async refreshMetadata(instancePath: string): Promise<void> {
     const stateManager = await this.app.registry.get(ServiceStateManager)
-    const state = stateManager.get<MutableState<ResourceState>>(getInstanceModStateKey(instancePath))
+    const state = stateManager.get<SharedState<ResourceState>>(getInstanceModStateKey(instancePath))
     if (state) {
       await state.revalidate()
       const modrinthClient = await this.app.registry.getOrCreate(ModrinthV2Client)
@@ -36,6 +36,7 @@ export class InstanceModsService extends AbstractService implements IInstanceMod
         try {
           const versions = await modrinthClient.getProjectVersionsByHash(all.map(v => v.hash))
           const options = Object.entries(versions).map(([hash, version]) => {
+            if (!hash) return undefined
             const f = all.find(f => f.hash === hash)
             if (f) return { hash: f.hash, metadata: { modrinth: { projectId: version.project_id, versionId: version.id } } }
             return undefined
@@ -51,6 +52,7 @@ export class InstanceModsService extends AbstractService implements IInstanceMod
         try {
           const chunkSize = 8
           const allChunks = [] as Resource[][]
+          all = all.filter(a => !!a.hash && !a.isDirectory)
           for (let i = 0; i < all.length; i += chunkSize) {
             allChunks.push(all.slice(i, i + chunkSize))
           }
@@ -60,13 +62,18 @@ export class InstanceModsService extends AbstractService implements IInstanceMod
             const prints = (await Promise.all(chunk.map(async (v) => ({ fingerprint: await worker.fingerprint(v.path), file: v }))))
             for (const { fingerprint, file } of prints) {
               if (fingerprint in allPrints) {
-                this.error(new Error(`Duplicated fingerprint ${fingerprint} for ${file.path} and ${allPrints[fingerprint].path}`))
+                this.warn(new Error(`Duplicated fingerprint ${fingerprint} for ${file.path} and ${allPrints[fingerprint].path}`))
                 continue
               }
               allPrints[fingerprint] = file
             }
           }
-          const result = await curseforgeClient.getFingerprintsMatchesByGameId(432, Object.keys(allPrints).map(v => parseInt(v, 10)))
+          const result = await curseforgeClient.getFingerprintsMatchesByGameId(432, Object.keys(allPrints).map(v => parseInt(v, 10))).catch((e) => {
+            if (e instanceof CurseforgeApiError && e.status >= 400 && e.status < 500 && e.status !== 404) {
+              this.error(e)
+            }
+            return { exactMatches: [] }
+          })
           const options = [] as { hash: string; metadata: { curseforge: { projectId: number; fileId: number } } }[]
           for (const f of result.exactMatches) {
             const r = allPrints[f.file.fileFingerprint] || Object.values(allPrints).find(v => v.hash === f.file.hashes.find(a => a.algo === 1)?.value)
@@ -110,15 +117,15 @@ export class InstanceModsService extends AbstractService implements IInstanceMod
     await this.app.shell.openDirectory(join(path, 'mods'))
   }
 
-  async watch(instancePath: string): Promise<MutableState<ResourceState>> {
+  async watch(instancePath: string): Promise<SharedState<ResourceState>> {
     if (!instancePath) throw new AnyError('WatchModError', 'Cannot watch instance mods on empty path')
-    const lock = this.semaphoreManager.getLock(LockKey.instance(instancePath))
+    const lock = this.mutex.of(LockKey.instance(instancePath))
     const stateManager = await this.app.registry.get(ServiceStateManager)
     return stateManager.registerOrGet(getInstanceModStateKey(instancePath), async ({ doAsyncOperation }) => {
       const basePath = join(instancePath, 'mods')
 
       await ensureDir(basePath)
-      const { dispose, revalidate, state } = this.resourceManager.watch(basePath, ResourceDomain.Mods, (func) => doAsyncOperation(lock.read(func)))
+      const { dispose, revalidate, state } = this.resourceManager.watch(basePath, ResourceDomain.Mods, (func) => doAsyncOperation(lock.waitForUnlock().then(func)))
 
       const instanceService = await this.app.registry.get(InstanceService)
       instanceService.registerRemoveHandler(instancePath, dispose)
@@ -131,9 +138,10 @@ export class InstanceModsService extends AbstractService implements IInstanceMod
 
   async installFromMarket(options: InstallMarketOptionWithInstance): Promise<string[]> {
     const provider = await this.app.registry.get(kMarketProvider)
-    const result = await provider.installFile({
+    const result = await provider.installInstanceFile({
       ...options,
-      directory: join(options.instancePath, 'mods'),
+      instancePath: options.instancePath,
+      domain: ResourceDomain.Mods,
     })
     return result.map(v => v.path)
   }
@@ -141,20 +149,31 @@ export class InstanceModsService extends AbstractService implements IInstanceMod
   async install({ mods: resources, path }: InstallModsOptions) {
     const promises: Promise<void>[] = []
     this.log(`Install ${resources.length} to ${path}/mods`)
+    const modDir = join(path, ResourceDomain.Mods)
     for (const res of resources) {
+      if (res.startsWith(modDir)) {
+        // some stupid case
+        continue
+      }
       const src = res
-      const dest = join(path, ResourceDomain.Mods, basename(res))
+      const dest = join(modDir, basename(res))
       const [srcStat, destStat] = await Promise.all([stat(src), stat(dest).catch(() => undefined)])
 
       let promise: Promise<any> | undefined
       if (!destStat) {
-        promise = linkWithTimeoutOrCopy(src, dest)
+        if (srcStat.isDirectory()) {
+          promise = linkDirectory(src, dest, this.logger)
+        } else {
+          promise = linkWithTimeoutOrCopy(src, dest)
+        }
       } else if (srcStat.ino !== destStat.ino) {
         promise = unlink(dest).then(() => linkWithTimeoutOrCopy(src, dest))
       }
       if (promise) {
         promises.push(promise.catch((e) => {
-          this.error(new Error(`Cannot deploy the resource from ${src} to ${dest}`, { cause: e }))
+          if (e.name === 'Error') {
+            e.name = 'ModInstallError'
+          }
           throw e
         }))
       }
